A highly optimized version has been added
use timer0. It is used in CTC (Clear Timer on Compare Match Mode) every
6ms, as the maximum pulse i approximately 5ms. As the Panasonic
protocol is very long (48 bits) is not necessary to check the length of
the pulses very precisely, as my other IR remotes protocols' are only
maximum 32 bits long. But it wouldn't be that hard to modify the code
to work with any IR protocol.